Configure disable diagnostic keymap
In some cases diagnostics are not helpful. For example when working with
Helm charts which are YAML documents with Go templates. Since in general
they are not valid YAML documents, LSP shows false errors which make
code less readable. In such cases it is better to just turn the
diagnostics off.